map section renders upon login
-moved map section to main HTML doc to prevent issues with jquery click
handlers embedded in map.
-added zoomStep to fix bug in map library
-rendering map on page load but hidden until log in.

notes: map data is hard coded for now. need to get countries/years from
database.
