+++ assets/scripts/templates/mapview.handlebars
@@ -1,1 +1,11 @@
-<div id="world-map" style="width: auto; height: 400px">os this rendering?</div>
+<!-- <section>
+  <a id='logout'>log-out</a>
+  <p>Some quote about adventure or something</p>
+  <button id='add-vacation'>Add an Adventure</button>
+<div id="world-map" style="width: 70%; height: 400px;"></div>
+<p>look back on your adventures</p>
+<form>
+  <input type=text>
+  <button type=submit>Go!</button>
+</form>
+</section> -->
