+++ assets/styles/index.scss
@@ -1,4 +1,4 @@
 $icon-font-path: '~bootstrap-sass/assets/fonts/bootstrap/';
 @import '~bootstrap-sass/assets/stylesheets/bootstrap';
-// @import './theme';
-// @import './map.css';
+@import './theme';
+@import './map.css';
