+++ assets/styles/map.css
@@ -1,6 +1,6 @@
 
 svg {
-  background: url("parchment-bg.png");
+  /*background: url("parchment-bg.png");*/
   /*#c3dbd1;*/
   touch-action: none;
 }
