+++ assets/styles/theme.scss
@@ -1,12 +1,11 @@
 /*body {
   background: #b5d3d9;
 }*/
-button {
-  background: rgba(225, 244, 244, 0.0);
-  border: 2px dashed #937455;
-  border-radius: 0px;
-  height: 25px;
-  color: white;
-  cursor: pointer;
-  margin: 5px;
-}
+// button {
+//   background: rgba(225, 244, 244, 0.0);
+//   border: 2px dashed #937455;
+//   border-radius: 0px;
+//   height: 25px;
+//   cursor: pointer;
+//   margin: 5px;
+// }
