+++ index.html
@@ -12,7 +12,19 @@
       <script src="public/application.js" type="text/javascript" charset="utf-8" defer></script>
     </head>
     <body class="container-fluid">
+      <h1>Travelers Keep</h1>
       <p id="section-alerts" class='alert'></p>
       <div id='content-container'></div>
+      <section id='map-view'>
+          <a id='logout'>log-out</a>
+          <p>Some quote about adventure or something</p>
+          <button id='add-vacation'>Add an Adventure</button>
+        <div id="world-map" style="width: auto; height: 400px;"></div>
+        <span>look back on your adventures</span>
+        <form>
+          <input type=text>
+          <button type=submit>Go!</button>
+        </form>
+      </section>
     </body>
 </html>
